First Push
The app is to present the current sentiment analysis on selected companies. The app scrapes live data (news) from a stock exchange site which is analyzed using nltk.vader. The sentiment analysis result is plotted to show both daily trends.
The app is for academic purposes only and should not be used singularly in determining an investment decision.

Areas of improvement on the app will be to gather more days data and to actually superimpose the plot of price movement to show the effect on the sentiments on the stock price.